Update Jenkinsfile
Insert the Build and Push Image stage after the existing Test stage and note the beforeAgent true option - this setting will result in the when condition being evaluated before acquiring and entering an agent for the stage. The branch condition is a built-in condition that allows executing stages only for specific branches - in this case the Build and Push Image stage will only execute for the master branch.